requirement.txt
langchain_groq
streamlit
langchain_community
pypdf
sentence-transformers